Integrate collection strawman scalacheck tests into main test suite
Like in the case of the JUnit tests, most tests in collection-strawman
were based on existing tests in scala/scala, adapted to the new package
structure, so there are very few actual changes to merge back.
