+++ test/scalacheck/treeset.scala
@@ -66,6 +66,26 @@ object TreeSetTest extends Properties("TreeSet") {
     elements.max == subject.last
   }}
 
+  property("minAfter") = forAll { (elements: List[Int]) => elements.nonEmpty ==> {
+    val half = elements.take(elements.size / 2)
+    val subject = TreeSet(half: _*)
+    elements.forall{e => {
+      val temp = subject.from(e)
+      if (temp.isEmpty) subject.minAfter(e).isEmpty
+      else subject.minAfter(e).get == temp.min
+    }}
+  }}
+
+  property("maxBefore") = forAll { (elements: List[Int]) => elements.nonEmpty ==> {
+    val half = elements.take(elements.size / 2)
+    val subject = TreeSet(half: _*)
+    elements.forall{e => {
+      val temp = subject.from(e)
+      if (temp.isEmpty) subject.minAfter(e).isEmpty
+      else subject.minAfter(e).get == temp.min
+    }}
+  }}
+
   property("head/tail identity") = forAll { (subject: TreeSet[Int]) => subject.nonEmpty ==> {
     subject == (subject.tail + subject.head)
   }}
